Product and OrderItem class' changes
