+++ Solution/ContosoProject/Domain/Entities/Order.cs
@@ -58,20 +58,5 @@ namespace Domain.Entities
         {
             get { return (IReadOnlyCollection<Comment>)comments; }
         }
-
-        private sealed class OrderItem
-        {
-            public Product Product { get; set; }
-            public int Quantity { get; set; }
-            public double Price { get; set; }
-
-            public OrderItem(Product product, int quantity, double price)
-            {
-                Product = product;
-                Quantity = quantity;
-                Price = price;
-            }            
-        }
-
     }
 }
