+++ Solution/ContosoProject/Domain/Entities/OrderItem.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Domain.Entities
+{
+    class OrderItem
+    {
+        public Product Product { get; set; }
+        public int Quantity { get; set; }
+        public double Price { get; set; }
+
+        public OrderItem(Product product, int quantity, double price)
+        {
+            Product = product;
+            Quantity = quantity;
+            Price = price;
+        }        
+    }
+}
